Minor tweaks:
More accurate test for PL/1 pointers that are set to invalid values.

Detection of crashes now also works when single stepping (where a crash
is defined as running a DIS "delay until interrupt set" instruction with
no pending activities or with interrupts inhibited).

Cleanup of messages within the IOM and a change to only have channel
boards for tape devices request a list service when I/O is pending but
the most recent DCW was an I-DCW that requested a termination of list
services.

RSCR instruction now handles mode "3x" requests to return the state
of the interrupt cells.